1)  Using File->Exit did not quit the app I have fixed this.     2)  I normally have the shift key mapped to rotate, control mapped to pan, and alt mapped to zoom.  Only the shift (rotate) key seemed to work. I have done this. Control was mapped to MultiSelect, what do you want mapped to MultiSelect?
I have got it to store the layout of the tool bars and the size of the window and I have added a default setting for it. I have had to add the tool bars at the beginning; previously they were being added when a new document was made.

Also, I have got wxWidgets to turn the toolbar images up the right way. This was easy; wxImage has a Mirror function.